+++ src/index.ts
@@ -40,3 +40,4 @@ export {
 } from './credentialDeploymentTransactions';
 export { isAlias, getAlias } from './alias';
 export { deserializeContractState } from './deserialization';
+export * from './blockSummaryHelpers';