+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.7.0 2022-03-09
+
+### Added
+
+- Support for getting baker list from node.
+- Support for getting status of Baker-/L-pool (required node to have protocol version 4 or later)
+- Helper functions for determining the version of `BlockSummary` and nested types. 
+- Support for initiating and updating contracts with parameters.
+
+### Changed
+
+- Updated `BlockSummary` type to include new version, effective from protocol version 4.
+- Updated `AccountInfo` type to include new fields related to delegation introduced with protocol version 4.
+
 ## 0.6.0 2022-02-02
 
 ### Added